requirements.txt
accelerate==1.11.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
fastapi==0.120.4
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
orjson==3.10.12
packaging==25.0
pillow==11.3.0
psutil==7.1.3
pydantic==2.12.3
pydantic_core==2.41.4
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
setuptools==70.2.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn[standard]==0.38.0
